Implemented sensors query, moto enable outputs and I2C io.

+++ soft/firmware/moto1/src/read_ctrl.h
@@ -5,7 +5,7 @@
 #include "ch.h"
 
 void     initRead( void );
-uint32_t valueRead( void );
+uint16_t valueRead( void );
 
 
 #endif
